SXDEDPCXZIC-90_DATAVIC-470 / move harvesters
