+++ ckanext/datavic_harvester/harvesters/__init__.py
@@ -1,3 +1,4 @@
-from ckanext.datavic_harvester.harvesters._json import DataVicDCATJSONHarvester
+from ckanext.datavic_harvester.harvesters.dcat_json import DataVicDCATJSONHarvester
 from ckanext.datavic_harvester.harvesters.metashare import MetaShareHarvester
 from ckanext.datavic_harvester.harvesters.delwp import DelwpHarvester
+from ckanext.datavic_harvester.harvesters.datavic_ckan import DataVicCKANHarvester
